Added and restructured tests
